+++ alembic/versions/ca90d428fa17_migration.py
@@ -1,0 +1,26 @@
+"""migration
+
+Revision ID: ca90d428fa17
+Revises: 
+Create Date: 2023-10-28 20:24:45.618442
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = 'ca90d428fa17'
+down_revision: Union[str, None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    pass
+
+
+def downgrade() -> None:
+    pass
